fixing problems in UI

+++ ariesframework/src/main/java/org/hyperledger/ariesframework/anoncreds/formats/anoncreds/EthrAnonCredsRegistry.kt
@@ -119,7 +119,7 @@ class EthrAnonCredsRegistry(override val methodName: String = "ethr") : AnonCred
             val revocationStatusList: RevocationStatusList =
                 agent.ledgerService.getRevocationStatusList(
                     revocationRegistryId,
-                    timestamp.toULong(),
+                    timestamp,
                 )
 
             val revocationStatusListInts: List<Int> =
